pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "AnthonyLoche", email = "anthonylocheifc@gmail.com"},
]
dependencies = [
    "django>=5.0.3",
    "djangorestframework>=3.15.0",
    "django-cors-headers>=4.3.1",
    "dotenv>=0.0.5",
    "djangorestframework-simplejwt>=5.3.1",
    "setuptools>=69.2.0",
    "python-magic>=0.4.27",
    "Pillow>=10.2.0",
    "libmagic>=1.0",
    "drf-spectacular>=0.27.1",
    "python-dotenv>=1.0.1",
    "netifaces>=0.11.0",
    "gunicorn>=21.2.0",
    "whitenoise[brotli]>=6.6.0",
    "psycopg2-binary>=2.9.9",
    "django-cloudinary-storage>=0.3.0",
    "django-filter>=24.2",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
pre_dev = "python ./scripts/set_my_ip.py"
dev = "python manage.py runserver 0.0.0.0:19003"